refactor: extract WorkoutTimeline from WorkoutTracker (#745)
Props+emits child (exercises/prBaselineDate/warmupThreshold in;
log-set/edit-set/delete-set out) following the RestTimerContent /
WorkoutCompleteView extraction pattern. Owns its session-local view
state (hideWarmups, pagination limit, tap-to-reveal actions). DOM and
classes are unchanged, so the existing timeline tests cover it through
the parent. The preferences-store test mock gains the filters field the
real store has always had — the template now reads it eagerly as a prop.

WorkoutTracker: 3,071 -> ~2,880 lines.
